Fix memdraw scanline stride on LP64 hosts.
Memimage.width is in 32-bit words; use sizeof(u32) instead of sizeof(ulong) so load/unload no longer overrun image pool data on arm64.

+++ libmemdraw/line.c
@@ -47,7 +47,7 @@ horline1(Memimage *dst, Point p0, Point p1, int srcval, Rectangle clipr)
 
 	deltax = p1.x - p0.x;
 	deltay = p1.y - p0.y;
-	dd = dst->width*sizeof(ulong);
+	dd = dst->width*sizeof(u32);
 	dy = 1;
 	if(deltay < 0){
 		dd = -dd;
@@ -114,7 +114,7 @@ verline1(Memimage *dst, Point p0, Point p1, int srcval, Rectangle clipr)
 			e += deltay;
 		}else
 			e += deltax;
-		d += dst->width*sizeof(ulong);
+		d += dst->width*sizeof(u32);
 		m >>= bpp;
 		if(m == 0)
 			m = m0;
